+++ glotaran/project/result.py
@@ -13,6 +13,7 @@ import xarray as xr
 from numpy.typing import ArrayLike
 from tabulate import tabulate
 
+from glotaran.deprecation import deprecate
 from glotaran.io import SAVING_OPTIONS_DEFAULT
 from glotaran.io import SavingOptions
 from glotaran.io import load_result
@@ -112,8 +113,8 @@ class Result:
 
     See also: :func:`scipy.optimize.least_squares`
     """
-    number_of_data_points: int | None = None
-    """Number of data points :math:`N`."""
+    number_of_residuals: int | None = None
+    """Number of values in the residual vector :math:`N`."""
     number_of_jacobian_evaluations: int | None = None
     """The number of jacobian evaluations."""
     number_of_parameters: int | None = None
@@ -143,6 +144,25 @@ class Result:
         if isinstance(self.jacobian, list):
             self.jacobian = np.array(self.jacobian)
             self.covariance_matrix = np.array(self.covariance_matrix)
+
+    @property
+    @deprecate(
+        deprecated_qual_name_usage="glotaran.project.Result.number_of_data_points",
+        new_qual_name_usage="glotaran.project.Result.number_of_residuals",
+        to_be_removed_in_version="0.8.0",
+        importable_indices=(2, 2),
+    )
+    def number_of_data_points(self) -> int | None:
+        """Return the number of values in the residual vector :math:`N`.
+
+        Deprecated since it returned the wrong value.
+
+        Returns
+        -------
+        int | None
+            Number of values in the residual vector :math:`N`.
+        """
+        return self.number_of_residuals
 
     @property
     def model(self) -> Model:
@@ -199,7 +219,7 @@ class Result:
         general_table_rows: list[list[Any]] = [
             ["Number of residual evaluation", self.number_of_function_evaluations],
             ["Number of parameters", self.number_of_parameters],
-            ["Number of datapoints", self.number_of_data_points],
+            ["Number of residuals", self.number_of_residuals],
             ["Degrees of freedom", self.degrees_of_freedom],
             ["Chi Square", f"{self.chi_square or np.nan:.2e}"],
             ["Reduced Chi Square", f"{self.reduced_chi_square or np.nan:.2e}"],