+++ changelog.md
@@ -51,6 +51,7 @@
 - Command Line Interface (removed without replacement) (#1228)
 - `Project.generate_model` (removed without replacement)
 - `Project.generate_parameters` (removed without replacement)
+- `glotaran.project.Result.number_of_data_points` -> `glotaran.project.Result.number_of_residuals`
 
 ### 🗑️❌ Deprecated functionality removed in this release